pyproject.toml
[tool.poetry]
name = "todo-api"
version = "0.1.0"
description = ""
authors = ["Linh Vo Duy <duylinh158@gmail.com"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
pymysql = "^1.1.1"
alembic = "^1.13.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pyjwt = "^2.8.0"
pydantic-settings = "^2.3.4"
pytz = "^2024.1"
python-multipart = "^0.0.9"
pydantic-i18n = "^0.4.4"
python-i18n = "^0.3.9"
loguru = "^0.7.2"
passlib = "^1.7.4"
bcrypt = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
